requirements.txt
gradio==3.47.1
torch==2.5.1
torchvision==0.18.1
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==9.5.0
